requirements.txt
flask
gunicorn
python-dateutil==2.8.1
beautifulsoup4==4.9.3
ScalaFunctional==0.6.0
requests==2.24.0
PyFunctional==1.4.2
pytest==6.1.1
google-cloud-storage==1.32.0